requirements.txt
django~=3.1.1
djangorestframework==3.12.1
django-cors-headers==3.5.0
mysqlclient==2.0.1
pymysql==0.10.1
node == 12.18.3
yarn == 1.22.10
react == 17.0.1
react-router-dom == 5.2.0